P1423R3 char8_t remediation: Refactor std::filesystem::path helper classes and functions.
This patch moves helper classes and functions for std::filesystem::path out of
the class definition to a detail namespace so that they are available to the
implementations of std::filesystem::u8path.  Prior to this patch, the SFINAE
constraints for those implementations were specified via delegation to the
overloads of path with a std::locale parameter; it just so happened that those
overloads had the same constraints.  As of P1423R3, u8path and those overloads
no longer have the same constraints, so this dependency must be broken.

This patch does not provide any intentional behavioral changes.