P0482R6 char8_t: declare std::c8rtomb and std::mbrtoc8 if provided by the C library.
This change completes implementation of the C++20 proposal P0482R6 by
adding declarations of std::c8rtomb() and std::mbrtoc8() if provided
by the C library.

Autoconf changes determine if the C library declares c8rtomb and mbrtoc8
at global scope when uchar.h is included and compiled with -fchar8_t;
_GLIBCXX_USE_UCHAR_CHAR8_T is defined if so.  The <cuchar> header
re-declares these functions in the std namespace only if available and
the C++20 __cpp_char8_t feature test macro is defined.

A new test is provided.  The test passes trivially if the C library does
not provide these functions.  Otherwise it ensures that the functions
are declared when <cuchar> is included and -fchar8_t support is enabled.