Prototype support for builtin types for wide and UTF character types in C.
This change provides command line options to enable use of builtin types for
wide and UTF character types in C code. Support for wchar_t is enabled by
the -fbuiltin-wchar_t option and indicated by the presence of a new
__BUILTIN_WCHAR_T__ predefined macro. Support for char8_t, char16_t, and
char32_t is enabled by the -fbuiltin-charn_t option and indicated by the
presence of a new __BUILTIN_CHARN_T__ macro.

Updates to wchar.h and uchar.t will be required to condition the typedefs
of the wide and UTF character types on the predefined macros noted above.